resolve #4
The problem is that the Base64 in JWT token is being compressed, which means there is no padding on the encoded string. That will cause error when trying to decode it using `atob()`.

Resolve that by adding a [function](https://github.com/Kunniii/edn_auto_ext/blob/v2/src/ready.js#L48) named `addPaddingToBase64()` and use it to add padding before converting it to a string.